Fix compilation and test failures

+++ generator-plugin-java8/src/test/java/org/stjs/generator/plugin/java8/writer/methodref/MethodReferenceGeneratorTest.java
@@ -10,20 +10,20 @@ public class MethodReferenceGeneratorTest {
 	@Test
 	public void testStaticMethodRef() {
 		assertCodeContains(MethodRef1.class, "calculate(MethodRef1.inc)");
-		assertEquals(Integer.valueOf(1), execute(MethodRef1.class));
+		assertEquals(Long.valueOf(1), execute(MethodRef1.class));
 	}
 
 	@Test
 	public void testInstanceMethodRef() {
 		assertCodeContains(MethodRef2.class,
 				"calculate(function(){return MethodRef2.prototype.inc2.call(arguments[0], arguments[1]);}, new MethodRef2(), 1)");
-		assertEquals(Integer.valueOf(3), execute(MethodRef2.class));
+		assertEquals(Double.valueOf(3.0), execute(MethodRef2.class));
 	}
 
 	@Test
 	public void testInstanceWithTargetMethodRef() {
 		assertCodeContains(MethodRef3.class, "calculate(stjs.bind(ref, \"inc2\"), 1)");
-		assertEquals(Integer.valueOf(4), execute(MethodRef3.class));
+		assertEquals(Double.valueOf(4.0), execute(MethodRef3.class));
 	}
 
 	@Test
